requirements.txt
python-dotenv==1.0.0
openai==1.3.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pygame==2.5.2
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0 